move config.js back to config/
